implement function of move tool: copy

+++ electron/app/actions/tool.ts
@@ -16,6 +16,7 @@ export const ADD_NOTE = 'ADD_NOTE';
 export const REMOVE_NOTE = 'REMOVE_NOTE';
 export const CLEAR_SELECTIONS = 'CLEAR_SELECTIONS';
 export const SET_SELECTIONS = 'SET_SELECTIONS';
+export const SET_KEEP_SELECTIONS = 'SET_KEEP_SELECTIONS';
 
 export function clearSelections() {
   return {
@@ -63,7 +64,7 @@ export function handleTool(toolType: string) {
   };
 }
 
-function addNote(note: MidiNote) {
+export function addNote(note: MidiNote) {
   return {
     type: ADD_NOTE,
     payload: {
@@ -74,11 +75,22 @@ function addNote(note: MidiNote) {
   };
 }
 
-function removeNote(note: MidiNote) {
+export function removeNote(note: MidiNote) {
   return {
     type: REMOVE_NOTE,
     payload: {
       note
+    },
+    meta: {},
+    error: null
+  };
+}
+
+export function setKeepSelections(value: boolean) {
+  return {
+    type: SET_KEEP_SELECTIONS,
+    payload: {
+      value
     },
     meta: {},
     error: null
@@ -101,7 +113,7 @@ function handleMouseRelease(
       });
     } else if (tool instanceof MoveTool) {
       selections.get('notes').forEach((note: MidiNote) => {
-        dispatch(removeNote(note));
+        if (!tool.get('shouldKeepSelection')) dispatch(removeNote(note));
       });
     } else if (tool instanceof LengthTool) {
       selections.get('notes').forEach((note: MidiNote) => {
@@ -149,126 +161,6 @@ export function onMouseEvent(
         error: new Error(`${event} is not defined.`)
       };
   }
-}
-
-export function onKeyEvent(event) {
-  return (dispatch: Dispatch, getState: GetState) => {
-    const { editor, preferences } = getState();
-    const tool = editor.get('tool');
-    const actionNames: Array<string> = [
-      'copy',
-      'paste',
-      'pitchUp',
-      'pitchDown',
-      'increaseBeat',
-      'decreaseBeat',
-      'removeNotes'
-    ];
-
-    let matchedAction = '';
-    actionNames.forEach((actionName: string) => {
-      const keyBind: KeyBind = preferences.keyBinds.get(actionName);
-      if (!keyBind) {
-        throw new Error(`${actionName}: is not defined in KeyBinds`);
-      }
-      if (
-        keyBind.codes.some((value: number) => {
-          return value === event.keyCode;
-        }) &&
-        keyBind.metaKey === event.metaKey &&
-        keyBind.altKey === event.altKey &&
-        keyBind.shiftKey === event.shiftKey
-      ) {
-        matchedAction = actionName;
-      }
-    });
-
-    console.log(matchedAction);
-
-    switch (matchedAction) {
-      case 'pitchUp':
-        {
-          const selections = tool.get('selections');
-          if (!tool.get('isDrawing')) {
-            selections.get('notes').forEach((note: MidiNote) => {
-              dispatch(removeNote(note));
-            });
-            selections.get('notes').forEach((note: MidiNote) => {
-              dispatch(
-                addNote(note.set('noteNumber', note.get('noteNumber') + 1))
-              );
-            });
-            dispatch(setSelections(selections.slidePitch(1)));
-          }
-        }
-        break;
-      case 'pitchDown':
-        {
-          const selections = tool.get('selections');
-          if (!tool.get('isDrawing')) {
-            selections.get('notes').forEach((note: MidiNote) => {
-              dispatch(removeNote(note));
-            });
-            selections.get('notes').forEach((note: MidiNote) => {
-              dispatch(
-                addNote(note.set('noteNumber', note.get('noteNumber') - 1))
-              );
-            });
-            dispatch(setSelections(selections.slidePitch(-1)));
-          }
-        }
-        break;
-      case 'increaseBeat':
-        {
-          const selections = tool.get('selections');
-          if (!tool.get('isDrawing')) {
-            selections.get('notes').forEach((note: MidiNote) => {
-              dispatch(removeNote(note));
-            });
-            selections.get('notes').forEach((note: MidiNote) => {
-              dispatch(
-                addNote(note.set('startBeat', note.get('startBeat') + 0.25))
-              );
-            });
-            dispatch(setSelections(selections.slideBeat(0.25)));
-          }
-        }
-        break;
-      case 'decreaseBeat':
-        {
-          const selections = tool.get('selections');
-          if (!tool.get('isDrawing')) {
-            selections.get('notes').forEach((note: MidiNote) => {
-              dispatch(removeNote(note));
-            });
-            selections.get('notes').forEach((note: MidiNote) => {
-              dispatch(
-                addNote(note.set('startBeat', note.get('startBeat') - 0.25))
-              );
-            });
-            dispatch(setSelections(selections.slideBeat(-0.25)));
-          }
-        }
-        break;
-      case 'removeNotes': {
-        const selections = tool.get('selections');
-        if (!tool.get('isDrawing')) {
-          selections.get('notes').forEach((note: MidiNote) => {
-            dispatch(removeNote(note));
-          });
-          dispatch(clearSelections());
-        }
-        break;
-      }
-      default:
-        return {
-          type: 'ERROR',
-          payload: {},
-          meta: {},
-          error: new Error(`${matchedAction} is not defined.`)
-        };
-    }
-  };
 }
 
 export function focusClip(trackId: number, clipId: number) {
